Description
# Simple-linear-Regression-in-ML
Simple Linear Regression

In Dataset there is 2 column and problem here we have 10 observations (Years of Experience  and Salary), 
and they hire you to find out if there is any sort of correlation or what is the correlation between salary and years of experience.
“This is one of the cases when management obviously knows that there is a correlation that they set the salaries”
What we are going to do in this  first of all you’re going to investigate what is the correlation.  
Here above it is mention that “The higher the years of exp is the higher the salary will be”. So, we will be creating a model a simple linear regression 
models which will tell us what is fitting line for this relationship. If you want to look at it from a business value add perspective what is us the business 
value out of this project apart from us if course learning about simple the regression the business value add here is that if the company doesn’t know they don’t 
have a certain set of rules how to set salaries. You can then go back to them with this model and say this is how. We can solve this issue.